+++ lib/src/tags/div/div.dart
@@ -1,12 +1,11 @@
 import 'package:wed/src/component/component.dart';
-
-import '../component/component_base_props.dart';
+import 'package:wed/src/tags/div/div_props.dart';
 
 /// The Div class is a component that renders a div element.
 ///
 /// It's used as a base class for the more specific DivXyz classes, such as DivRow and DivColumn.
 ///
-class Div<T extends ComponentBaseProps> extends Component<T> {
+class Div extends Component<DivProps> {
   Div({
     super.key,
     super.ref,