+++ lib/src/tags/div/div_props.dart
@@ -1,0 +1,8 @@
+import '../../../wed.dart';
+
+class DivProps extends ComponentBaseProps {
+  const DivProps({
+    super.styles,
+    super.children,
+  });
+}