+++ lib/wed.dart
@@ -1,7 +1,19 @@
 library wed;
 
+// component
 export 'src/component/component.dart';
-export 'src/css/css_base_style.dart';
 export 'src/component/component_base_props.dart';
-export 'src/tags/div.dart';
+
+// css
+export 'src/css/css_base_style.dart';
+export 'src/css/props/css_prop.dart';
+export 'src/css/props/border_radius.dart';
+
+// tags
+export 'src/tags/div/div.dart';
+export 'src/tags/div/div_props.dart';
+export 'src/tags/button/button.dart';
+export 'src/tags/button/button_props.dart';
+
+// fundamentals
 export 'src/fundamentals.dart';