+++ web/main.dart
@@ -1,22 +1,11 @@
 import 'dart:html';
 
-import 'package:wed/src/tags/button/button.dart';
-import 'package:wed/src/tags/button/button_props.dart';
 import 'package:wed/wed.dart';
 
-final app = querySelector("#app") as DivElement;
+final app = querySelector("#output") as DivElement;
 
 void main() {
-  const styles = CssBaseStyle(
-    backgroundColor: 'red',
-    height: '100px',
-    width: '100px',
-    alignItems: 'center',
-    justifyContent: 'center',
-    display: 'flex',
-  );
-
-  final child = Button(
+  final button = Button(
     props: ButtonProps(
       text: 'Click me',
       onClick: () => print('Clicked'),
@@ -24,16 +13,40 @@ void main() {
         height: '50px',
         width: '150px',
         textAlign: 'center',
+        color: 'white',
+        backgroundColor: 'purple',
       ),
     ),
   );
 
+  final div = Div(
+    props: DivProps(
+      children: [button],
+      styles: CssBaseStyle(
+        height: '300px',
+        width: '300px',
+        display: 'flex',
+        alignItems: 'center',
+        justifyContent: 'center',
+        backgroundColor: 'white',
+        borderRadius: BorderRadius.all(50),
+      ),
+    ),
+  );
+
+  const componentStyle = CssBaseStyle(
+    height: '100vh',
+    width: '100vw',
+    backgroundColor: 'blue',
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+  );
+
   final component = Div(
-    key: '',
-    ref: '',
-    props: ComponentBaseProps(
-      styles: styles,
-      children: [child],
+    props: DivProps(
+      styles: componentStyle,
+      children: [div],
     ),
   );